requirements.txt
gradio
gradio-client
pillow
requests